Cast result of Enumeration to int

+++ Classes/InfoBox/AbstractInfoBox.php
@@ -83,7 +83,9 @@ abstract class AbstractInfoBox implements \SplObserver
         if ($subject instanceof Page) {
             $this->view->assign('body', $this->renderBody($subject));
             $this->view->assign('unorderedList', $this->unorderedList);
-            $this->view->assign('state', (string)$this->state);
+
+            // First call __string() of Enumeration, then cast to INT
+            $this->view->assign('state', (int)(string)$this->state);
 
             // $shouldBeRendered can be modified within renderBody() by a developer
             if ($this->shouldBeRendered) {
